chore: remove residual dead code and harden test gates

+++ modules/learn/validators/engagement.ts
@@ -9,5 +9,5 @@ export const BadgeNotifySchema = z.object({
 })
 
 export const XpHistoryQuerySchema = z.object({
-  limit: z.coerce.number().min(1).max(100).default(20),
+  limit: z.coerce.number().int().min(1).max(100).default(20),
 })
